Rename list_php_binaries to php_binaries
